abstract configuration file parse
