+++ framework/config/parse/parser.go
@@ -1,9 +1,15 @@
 package parse
 
-//func NewParser(cfgType string) Parser {
-//	if cfgType == "yaml" {
-//		return &
-//	}
-//
-//	return nil
-//}
+type CfgParser interface {
+	Unmarshal(in []byte, out interface{}) error
+	SupportFileType() []string
+}
+
+func NewCfgParser(cfgType string) CfgParser {
+	switch cfgType {
+	case "yaml":
+		return yamlParser{cfgType: cfgType}
+	default:
+		return nil
+	}
+}
